# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HirenextAI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: getJobs
      tags: [jobs]
      summary: Get job listings
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
        - name: remote
          in: query
          schema:
            type: boolean
        - name: type
          in: query
          schema:
            type: string
            enum: [full-time, part-time, internship, freelance]
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Job listings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobsResponse"
  /jobs/{jobId}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get a single job
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateProfile
      tags: [profile]
      summary: Update user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    Job:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        company:
          type: string
        companyLogoUrl:
          type: string
          nullable: true
        location:
          type: string
        isRemote:
          type: boolean
        type:
          type: string
          enum: [full-time, part-time, internship, freelance]
        category:
          type: string
        salaryMin:
          type: integer
          nullable: true
        salaryMax:
          type: integer
          nullable: true
        description:
          type: string
        requirements:
          type: array
          items:
            type: string
        skills:
          type: array
          items:
            type: string
        postedAt:
          type: string
          format: date-time
        isFresher:
          type: boolean
        experienceYears:
          type: integer
          nullable: true
        applicationCount:
          type: integer
        source:
          type: string
          description: "Source of the job listing (e.g. JSearch, Adzuna, Arbeitnow, Database)"
          nullable: true
        applyUrl:
          type: string
          nullable: true
          description: "External URL to apply for the job"
      required:
        - id
        - title
        - company
        - location
        - isRemote
        - type
        - category
        - description
        - requirements
        - skills
        - postedAt
        - isFresher
        - applicationCount
    JobsResponse:
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/Job"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - jobs
        - total
        - page
        - totalPages
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        headline:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        skills:
          type: array
          items:
            type: string
        education:
          type: array
          items:
            $ref: "#/components/schemas/Education"
        experience:
          type: array
          items:
            $ref: "#/components/schemas/Experience"
        resumeUrl:
          type: string
          nullable: true
        preferredLocations:
          type: array
          items:
            type: string
        preferredCategories:
          type: array
          items:
            type: string
        openToRemote:
          type: boolean
        expectedSalaryMin:
          type: integer
          nullable: true
        expectedSalaryMax:
          type: integer
          nullable: true
        isFresher:
          type: boolean
        degreeLevel:
          type: string
          nullable: true
        specialization:
          type: string
          nullable: true
        setupCompleted:
          type: boolean
        completionPct:
          type: integer
      required:
        - id
        - userId
        - skills
        - education
        - experience
        - preferredLocations
        - preferredCategories
        - openToRemote
        - isFresher
        - setupCompleted
        - completionPct
    Education:
      type: object
      properties:
        institution:
          type: string
        degree:
          type: string
        field:
          type: string
        startYear:
          type: integer
        endYear:
          type: integer
          nullable: true
        current:
          type: boolean
      required:
        - institution
        - degree
        - field
        - startYear
        - current
    Experience:
      type: object
      properties:
        company:
          type: string
        title:
          type: string
        location:
          type: string
          nullable: true
        startDate:
          type: string
        endDate:
          type: string
          nullable: true
        current:
          type: boolean
        description:
          type: string
          nullable: true
      required:
        - company
        - title
        - startDate
        - current
    UpdateProfileRequest:
      type: object
      properties:
        headline:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        skills:
          type: array
          items:
            type: string
        education:
          type: array
          items:
            $ref: "#/components/schemas/Education"
        experience:
          type: array
          items:
            $ref: "#/components/schemas/Experience"
        preferredLocations:
          type: array
          items:
            type: string
        preferredCategories:
          type: array
          items:
            type: string
        openToRemote:
          type: boolean
        expectedSalaryMin:
          type: integer
          nullable: true
        expectedSalaryMax:
          type: integer
          nullable: true
        isFresher:
          type: boolean
        degreeLevel:
          type: string
          nullable: true
        specialization:
          type: string
          nullable: true
        setupCompleted:
          type: boolean
...